# openapi.yaml
openapi: 3.0.3
info:
  title: Email Validation API
  description: API for validating email addresses in real-time or bulk
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
    description: Local development server
paths:
  /validate:
    post:
      summary: Validate a single email address
      description: Validates an email address through syntax, domain, MX, and SMTP checks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "user@example.com"
                options:
                  type: object
                  properties:
                    skip_smtp:
                      type: boolean
                      description: Skip SMTP verification for faster results
                      example: false
      responses:
        '200':
          description: Successful validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid request
        '500':
          description: Server error
  /validate/bulk:
    post:
      summary: Validate multiple email addresses
      description: Validates a batch of email addresses asynchronously
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emails:
                  type: array
                  items:
                    type: string
                  example: ["user1@example.com", "user2@example.com"]
                options:
                  type: object
                  properties:
                    skip_smtp:
                      type: boolean
                      description: Skip SMTP verification for faster results
                      example: false
      responses:
        '200':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                    example: "550e8400-e29b-41d4-a716-446655440000"
                  status:
                    type: string
                    example: "queued"
        '400':
          description: Invalid request
        '500':
          description: Server error
  /validate/bulk/csv:
    post:
      summary: Validate email addresses from CSV file
      description: Validates email addresses from a CSV file with self-mapping functionality
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csvFile:
                  type: string
                  format: binary
                  description: The CSV file containing email addresses
                emailColumn:
                  type: string
                  description: The name of the column containing email addresses
                  example: "email"
                options:
                  type: string
                  description: JSON string with validation options
                  example: '{"skip_smtp": true}'
      responses:
        '200':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                    example: "550e8400-e29b-41d4-a716-446655440000"
                  status:
                    type: string
                    example: "queued"
        '400':
          description: Invalid request
        '500':
          description: Server error
  /jobs/{jobId}:
    get:
      summary: Get job status
      description: Retrieve the status of a bulk validation job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found
        '500':
          description: Server error
  /jobs/{jobId}/results:
    get:
      summary: Get job results
      description: Retrieve the results of a completed bulk validation job in JSON format
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job results
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationResult'
        '404':
          description: Job not found
        '500':
          description: Server error
  /jobs/{jobId}/results/csv:
    get:
      summary: Download job results as CSV
      description: Download the results of a completed bulk validation job as a CSV file
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: CSV file with validation results
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '404':
          description: Job not found
        '400':
          description: Job not completed
        '500':
          description: Server error
components:
  schemas:
    ValidationResult:
      type: object
      properties:
        email:
          type: string
          example: "user@example.com"
        syntax:
          type: boolean
          example: true
        mx:
          type: array
          items:
            type: string
          example: ["mx1.example.com", "mx2.example.com"]
        smtp:
          type: object
          properties:
            ok:
              type: boolean
              example: true
            response:
              type: string
              example: "250 Accepted"
        status:
          type: string
          enum: [valid, invalid, risky, unknown]
          example: "valid"
        score:
          type: integer
          example: 95
        disposable:
          type: boolean
          example: false
        role:
          type: boolean
          example: false
    JobStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        total:
          type: integer
        completed:
          type: integer
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true